requirements.txt
flask==3.0.3
pandas==2.2.2
groq==0.9.0
python-dotenv==1.0.1
